requirements.txt
gradio
openai
openai[voice_helpers]
python-dotenv